💁 start porting EarthEnv to new API
